backend/requirements.txt
FastAPI
uvicorn
mediapipe
tensorflow-cpu
numpy
opencv-python-headless
scikit-learn
python-multipart
